fix: keep Playwright and Inkeep types compatible with pnpm
