+++ package.json
@@ -31,6 +31,7 @@
     "@fingerprint/react": "^3.0.0",
     "@fingerprintjs/fingerprintjs-pro-server-api": "^6.10.0",
     "@inkeep/cxkit-react": "^0.5.107",
+    "@inkeep/cxkit-types": "0.5.107",
     "@mui/icons-material": "^5.15.11",
     "@mui/material": "^5.15.11",
     "@radix-ui/react-accordion": "^1.1.2",
